+++ modules/database/connect.py
@@ -6,7 +6,7 @@ from typing import Optional, Type
 import os
 import apsw  # type: ignore
 from apsw import Error
-from modules.logger_utils import init_logger
+from modules.utils.log import init_logger
 
 logger = init_logger()
