+++ modules/feeds/domainsproject.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Tuple, Iterator
 from more_itertools.more import chunked, sort_together
 
 from modules.feeds.hostname_expressions import generate_hostname_expressions
-from modules.logger_utils import init_logger
+from modules.utils.log import init_logger
 
 logger = init_logger()
